sbverify: Read --cert parameter as a cert store
Make sure the --cert parameter when passed to sbverify will read the file
as a potential full certificate store (concatenated PEM certificates), as they
are trivially extracted from a signed binary or as they might be shipped
separately the context of binaries with multiple signatures.

This makes sure one can do sanity checks for the tools and build binaries such
as extracting the signatures to PKCS7, converting that back to PEM, and using
that certificates file directly with sbverify.

Signed-off-by: Mathieu Trudel-Lapierre <mathieu.trudel-lapierre@canonical.com>